Initial checkin.
* Setup testing framework.
* Create base jslinq.js file.
