+++ test-context.js
@@ -1,0 +1,5 @@
+require('babel-polyfill');
+
+var context = require.context('./tests', true, /-spec\.js$/);
+
+context.keys().forEach(context);
